requirements.txt
python-telegram-bot[job-queue]==20.7
python-dotenv==1.0.0
aiohttp==3.9.1